requirements.txt
langchain
wikipedia
langchain-community
langchain-anthropic
langchain-openai
langchain-ai21
python-dotenv
pydantic
duckduckgo-search
WikipediaQueryRun
